Clear filters prior to returning edited rows.
If the datatable is filtered, the function to return the edited collection returns incorrect results.  Clearing the filters prior to accessing these rows should not present much of a problem to the UI, as the only time these edited rows will be returned (well - in normal usage) will be when the data is going to be saved.  Thus, clearing the filters won't hurt, and will fix the problem of data being missed when sending to a POST or what have you.